Nav Tabs and Mobile Layout
Created visual effects and adjusted media margins for Nav Tabs. Restructured React strap column layout back into div containers, and adjusted css margins to fit desktop. Needed this to allow for a column stack of components in mobile responsiveness. Otherwise it was layered with one column over the top of the other. "